refactor: move blob to internal/pkg
